moved /pages to /src
