+++ src/pages/_app.tsx
@@ -2,9 +2,9 @@ import { AppProps } from "next/app";
 import React from "react";
 import { ThemeProvider } from "styled-components";
 
-import { Layout } from "../src/components/Layout";
-import { GlobalStyle } from "../src/design-system/globals/globalStyle";
-import { lightTheme } from "../src/design-system/theme/lightTheme";
+import { Layout } from "../components/Layout";
+import { GlobalStyle } from "../design-system/globals/globalStyle";
+import { lightTheme } from "../design-system/theme/lightTheme";
 
 const App: React.FC<AppProps> = ({ Component, pageProps }) => {
   return (
